remove PetitionProjectMixin from instructor view
