+++ envergo/petitions/services.py
@@ -1,15 +1,19 @@
 import logging
 from dataclasses import dataclass
 from textwrap import dedent
-from typing import Literal
+from typing import Any, List, Literal
+from urllib.parse import urlparse
 
 import requests
 from django.conf import settings
 from django.contrib.sites.models import Site
+from django.http import QueryDict
 from django.urls import reverse
 
 from envergo.moulinette.forms import MOTIF_CHOICES
+from envergo.moulinette.models import MoulinetteHaie
 from envergo.utils.mattermost import notify
+from envergo.utils.tools import display_form_details
 
 logger = logging.getLogger(__name__)
 
@@ -507,3 +511,240 @@ Requête envoyée :
     usager = (dossier.get("usager") or {}).get("email", "")
 
     return DemarchesSimplifieesDetails(applicant_name, city, pacage, usager)
+
+
+def get_moulinette_from_project(petition_project):
+    parsed_url = urlparse(petition_project.moulinette_url)
+    query_string = parsed_url.query
+    # We need to convert the query string to a flat dict
+    raw_data = QueryDict(query_string)
+    moulinette_data = raw_data.dict()
+    moulinette_data["haies"] = petition_project.hedge_data
+    moulinette = MoulinetteHaie(
+        moulinette_data,
+        moulinette_data,
+        False,
+    )
+    return moulinette
+
+
+class Alert:
+    def __init__(self, key, extra: dict[str, Any] = dict, is_fatal=False):
+        self.key = key
+        self.extra = extra
+        self.is_fatal = is_fatal
+
+
+class AlertList(List[Alert]):
+    def __init__(self, request):
+        super().__init__()
+        self.request = request
+        self._config = None
+        self._petition_project = None
+        self._form = None
+        self._user_error_reference = None
+
+    @property
+    def petition_project(self):
+        return self._petition_project
+
+    @petition_project.setter
+    def petition_project(self, value):
+        self._petition_project = value
+
+    @property
+    def config(self):
+        return self._config
+
+    @config.setter
+    def config(self, value):
+        self._config = value
+
+    @property
+    def form(self):
+        return self._form
+
+    @form.setter
+    def form(self, value):
+        self._form = value
+
+    @property
+    def user_error_reference(self):
+        return self._user_error_reference
+
+    @user_error_reference.setter
+    def user_error_reference(self, value):
+        self._user_error_reference = value
+
+    def append(self, item: Alert) -> None:
+        if not isinstance(item, Alert):
+            raise TypeError("Only Alert instances can be added to AlertList")
+        super().append(item)
+
+    def compute_message(self):
+        lines = []
+        config_url = None
+        if self.config:
+            config_relative_url = reverse(
+                "admin:moulinette_confighaie_change", args=[self.config.id]
+            )
+            config_url = self.request.build_absolute_uri(config_relative_url)
+
+        if self.object:
+            lines.append("#### Mapping avec Démarches-simplifiées : :warning: anomalie")
+            projet_relative_url = reverse(
+                "admin:petitions_petitionproject_change",
+                args=[self.object.id],
+            )
+            projet_url = self.request.build_absolute_uri(projet_relative_url)
+
+            lines.append("Un dossier a été créé sur démarches-simplifiées : ")
+            lines.append(f"* [projet dans l’admin]({projet_url})")
+
+            if self.config:
+                dossier_url = (
+                    f"https://www.demarches-simplifiees.fr/procedures/"
+                    f"{self.config.demarche_simplifiee_number}/dossiers/"
+                    f"{self.object.demarches_simplifiees_dossier_number}"
+                )
+                lines.append(
+                    f"* [dossier DS n°{self.object.demarches_simplifiees_dossier_number}]"
+                    f"({dossier_url}) (:icon-info:  le lien ne sera fonctionnel qu’après le dépôt du dossier"
+                    f" par le pétitionnaire)"
+                )
+            if config_url:
+                lines.append("")
+                lines.append(
+                    f"Une ou plusieurs anomalies ont été détectées dans la [configuration du département "
+                    f"{self.config.department}]({config_url})"
+                )
+
+            lines.append("")
+            lines.append(
+                "Le dossier a été créé sans encombres, mais il contient peut-être des réponses sans pré-remplissage "
+                "ou avec une valeur erronnée."
+            )
+
+        else:
+            lines.append("### Mapping avec Démarches-simplifiées : :x: erreur")
+
+            lines.append(
+                "La création d’un dossier démarches-simplifiées n’a pas pu aboutir."
+            )
+
+            if config_url:
+                lines.append(
+                    f"Cette erreur révèle une possible anomalie de la [configuration du département "
+                    f"{self.config.department}]({config_url})"
+                )
+
+            lines.append(
+                f"L’utilisateur a reçu un message d’erreur avec l’identifiant `{self.user_error_reference.upper()}` "
+                f"l’invitant à nous contacter."
+            )
+
+            if self.form:
+                lines.append(
+                    f"""
+* form :
+```
+{display_form_details(self.form)}
+```
+"""
+                )
+
+        index = 0
+
+        for alert in self:
+            index = index + 1
+            if alert.is_fatal:
+                lines.append("")
+                lines.append(f"#### :x: Description de l’erreur #{index}")
+            else:
+                lines.append("")
+                lines.append(f"#### :warning: Description de l’anomalie #{index}")
+
+            if alert.key == "missing_demarche_simplifiee_number":
+                lines.append(
+                    "Un département activé doit toujours avoir un numéro de démarche sur Démarches Simplifiées"
+                )
+
+            elif alert.key == "invalid_prefill_field":
+                lines.append(
+                    "Chaque entrée de la configuration de pré-remplissage doit obligatoirement avoir un id et une "
+                    "valeur. Le mapping est optionnel."
+                )
+                lines.append(f"* Champ : {alert.extra['field']}")
+
+            elif alert.key == "ds_api_http_error":
+                lines.append(
+                    "L'API de Démarches Simplifiées a retourné une erreur lors de la création du dossier."
+                )
+                lines.append(
+                    dedent(
+                        f"""
+                **Requête:**
+                * url : {alert.extra['api_url']}
+                * body :
+                ```
+                {alert.extra['request_body']}
+                ```
+
+                **Réponse:**
+                * status : {alert.extra['response'].status_code}
+                * content:
+                ```
+                {alert.extra['response'].text}
+                ```
+                """
+                    )
+                )
+
+            elif alert.key == "missing_source_regulation":
+                lines.append(
+                    f"La configuration demande de pré-remplir un champ avec la valeur de **{alert.extra['source']}** "
+                    f"mais la moulinette n'a pas de résultat pour la réglementation "
+                    f"**{alert.extra['regulation_slug']}**."
+                )
+
+            elif alert.key == "missing_source_criterion":
+                lines.append(
+                    f"La configuration demande de pré-remplir un champ avec la valeur de **{alert.extra['source']}** "
+                    f"mais la moulinette n'a pas de résultat pour le critère "
+                    f"**{alert.extra['criterion_slug']}**."
+                )
+
+            elif alert.key == "missing_source_moulinette":
+                lines.append(
+                    f"La configuration demande de pré-remplir un champ avec la valeur de **{alert.extra['source']}** "
+                    f"mais la simulation ne possède pas cette valeur."
+                )
+
+            elif alert.key == "mapping_missing_value":
+                lines.append(
+                    dedent(
+                        f"""\
+               Une valeur prise en entrée n’a pas été reconnue dans le mapping
+               * Champ : {alert.extra['source']}
+               * Valeur : {alert.extra['value']}
+               * mapping :
+               ```
+               {alert.extra['mapping']}
+               ```
+               """
+                    )
+                )
+
+            elif alert.key == "invalid_form":
+                lines.append("Le formulaire contient des erreurs")
+
+            elif alert.key == "unknown_error":
+                lines.append("Nous ne savons pas d'où provient ce problème...")
+
+            else:
+                logger.error(
+                    "Unknown alert key during petition project creation",
+                    extra={"alert": alert},
+                )
+
+        return "\n".join(lines)
